Finished Logger
Logger is the CCS project gyro-readings-logger. It works by sending UART to a terminal such as Tera-Term and then saving the console log. log.txt is ~30 minutes of runtime. When using Tera-term you can use BUSPIRATETERA.INI to access the bus pirate. During the process I edited uart.c and uart.h to reflect the updated versions in examples